requirements.txt
langgraph
langchain-openai
python-dotenv
flask
pdfplumber
python-docx
pillow
